feat(core): cleanup unused code and duplicate methods
