+++ packages/core/src/operations/index.ts
@@ -1,3 +1,1 @@
-export * from "./getText";
-export * from "./getObject";
-export * from "./getBoolean";
+export * from "./base";
