+++ packages/core/src/runtime/templates.ts
@@ -101,6 +101,27 @@ function isConversationMessage(value: unknown): value is ConversationMessage {
   );
 }
 
+// Helper to extract valid JSON from a string that might contain code blocks or extra text
+export function extractJson(str: string): string {
+  // Remove markdown code blocks
+  str = str.replace(/```(?:\w*\s*)\n?/g, "").replace(/```/g, "");
+
+  // Find the last occurrence of a JSON structure (after any thinking tags)
+  const matches = str.match(/\{[\s\S]*\}|\[[\s\S]*\]/g);
+  if (!matches) {
+    throw new Error("No JSON-like structure found in response");
+  }
+
+  // Return the last match (after any thinking/reasoning)
+  return matches[matches.length - 1] ?? "";
+}
+
+// Helper to clean JSON string before parsing
+export function cleanJsonString(str: string): string {
+  // Just trim whitespace - the model should be giving us valid JSON
+  return str.trim();
+}
+
 // Template sections
 const EXAMPLES_SECTION = `<examples>
